configs/releng/airootfs/etc/mkinitcpio.conf.d/archiso.conf: use xz -9e
Use higher compression for the initramfs.

Now that mkinitcpio moves the compressed kernel modules and firmware
files to the early uncompressed initramfs, we can compress the main
initramfs image slightly more without it increasing the built time too
much. This will increase the memory required for decompression from
9 MiB to 65 MiB (assuming the kernel's decompressor works the same as
the xz tool), but that should not be an issue for any system targeting
Arch.

This reduces the initramfs size by ~250 KiB, which will save ~500 KiB
for the whole ISO.