configs/baseline/profiledef.sh: do not use mkfs.erofs extended options "fragments" and "dedupe"
These options are not compatible with multi-threaded compression and
actually increase the image file size.

With this change, the baseline profile's EROFS image file size is
reduced from 488 MiB to 398 MiB.
On my system, the compression time is reduced from about 2m6.274s to
1m59.410s when using erofs-utils built without multi-threaded
compression and to 0m27.883s when using erofs-utils built with
multi-threaded compression.