mkarchiso: show the FAT image size in MiB
The size of `efiboot.img` is never a fraction of a mebibyte (MiB) since
it is always rounded up to the next mebibyte. That means there is no need
to show the size in kibibytes (KiB) since it is easier to comprehend the
value if it is presented in mebibytes.